+++ src/api/scope/lib/scope-init.js
@@ -1,6 +1,6 @@
 /** @flow */
 import { Scope } from '../../../scope';
 
-export default function init(path: string, name: string, groupName: string): Promise<Scope> {
+export default function init(path: string = process.cwd(), name: string, groupName: string): Promise<Scope> {
   return Scope.ensure(path, name, groupName).then(scope => scope.ensureDir());
 }